resume button and some nav styling
