+++ client/src/assets/index.js
@@ -1,0 +1,1 @@
+export { default as Resume } from './pdf/Krisi-Keranova.pdf';
