+++ client/src/assets/pdf/Krisi-Keranova.pdf
