+++ client/src/styling/Header.css
@@ -26,13 +26,15 @@
 }
 ul {
   list-style-type: none;
+  margin-top: 0;
+  margin-bottom: 0;
 }
 
 .menu a {
   color: whitesmoke;
   cursor: pointer;
   margin: 0;
-  width: 50%;
+  /* width: 50%; */
   height: 100%;
   padding: 15px;
   text-decoration: none;
